Move Weekday into a separate module
